Added LoggingMiddleWare and Configured for Docker and Render deployment

+++ Backend/Controllers/ConversationsController.cs
@@ -24,7 +24,12 @@ namespace Backend.Controllers
                 return BadRequest(new { message = "Invalid user id." });
             }
 
-            return await _conversationsService.GetAllConversationsAsync(userModel.UserId);
+
+            Console.WriteLine("Fetching all conversations...");
+            var response = await _conversationsService.GetAllConversationsAsync(userModel.UserId);
+
+            Console.WriteLine("Conversations fetched successfully.");
+            return response;
         }
 
         [HttpPost]
@@ -36,12 +41,15 @@ namespace Backend.Controllers
                 return BadRequest(new { message = "Invalid conversation id." });
             }
 
+            Console.WriteLine("Fetching conversation...");
             Conversation? conversation = await _conversationsService.GetConversationByIdAsync(conversationModel.Id);
             if (conversation == null)
             {
+                Console.WriteLine("Conversation not found.");
                 return NotFound(new { message = "Conversation not found" });
             }
 
+            Console.WriteLine("Conversation fetched successfully.");
             return Ok(conversation);
         }
 
@@ -49,12 +57,20 @@ namespace Backend.Controllers
         [Route("")]
         public async Task<IActionResult> DeleteConversation(ConversationModel conversationModel)
         {
+            if (string.IsNullOrWhiteSpace(conversationModel.Id))
+            {
+                return BadRequest(new { message = "Invalid conversation id." });
+            }
+
+            Console.WriteLine("Deleting conversation...");
             var deleted = await _conversationsService.DeleteConversationAsync(conversationModel.Id);
             if (!deleted)
             {
+                Console.WriteLine("Conversation not found.");
                 return NotFound(new { message = "Conversation not found." });
             }
 
+            Console.WriteLine("Conversation deleted successfully.");
             return Ok(new { message = "Conversation deleted successfully." });
         }
     }
